Initial implementation
A new box is added to GUI where the BPM multiplier can be provided. If it is provided, it is used to ensure that all the notes and gaps between notes are a multiple of the BPM multiplier. This makes mapping most songs easier, but it might mess up some songs where the notes are weird.