Add safeguards for loud input
With the previous settings, issues could appear when the input (kara.moe) audio was already loud. Therefore, the loudness range target for the EBU algorithm has also been maximised and the --keep-lra-above-loudness-range-target option has been added to ensure that the loudest settings will be used and that ffmpeg-normalize won't automatically adjust things in an unexpected way.