Use true bpm instead of multiplier
The term multiplier can be confusing as to the role of what it actually does etc. True BPM should be much more easily understandable, especially when contrasted with the term karaoke BPM, hopefully providing a better UX. The only issue is the requirement for the karaoke BPM to be an integer multiple of the true BPM, but that is enforced through an error message.