Fix syllable being cut off to different line
Sometimes, the first syllable would be moved to the previous line incorrectly. This was because the line break was being placed at the end of the last syllable of each line. However, if due to overlap the next line started before the previous line ended, the line break would end up being placed after the first syllable of the next line. This would result in all the syllables from the next line that start before the line break being placed into the previous line. This happened because the lines are sorted before being saved to disk. It has been fixed by placing the line break at the start of the next line's first syllable instead.